Feature/section-5

+++ src/items/items.service.ts
@@ -1,7 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common'
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException
+} from '@nestjs/common'
 import { Item } from '../entities/item.entity'
 import { CreateItemDto } from './dto/create-item.dto'
 import { ItemRepository } from './items.repository'
+import { User } from 'src/entities/user.entity'
 
 @Injectable()
 export class ItemsService {
@@ -20,18 +25,25 @@ export class ItemsService {
     return found
   }
 
-  async create(createItemDto: CreateItemDto): Promise<Item> {
-    return await this.itemRepository.createItem(createItemDto)
+  async create(createItemDto: CreateItemDto, user: User): Promise<Item> {
+    return await this.itemRepository.createItem(createItemDto, user)
   }
 
-  async updateStatus(id: string): Promise<Item> {
+  async updateStatus(id: string, user: User): Promise<Item> {
     const item = await this.findById(id)
+    if (item.userId === user.id) {
+      throw new BadRequestException('自身の商品を購入することはできません。')
+    }
     item.status = 'SOLD_OUT'
     item.updatedAt = new Date().toISOString()
     return await this.itemRepository.updateStatus(item)
   }
 
-  async deleteItem(id: string): Promise<void> {
+  async deleteItem(id: string, user: User): Promise<void> {
+    const item = await this.findById(id)
+    if (item.userId !== user.id) {
+      throw new BadRequestException('他人の商品を削除することはできません。')
+    }
     await this.itemRepository.delete(id)
   }
 }
